requirements.txt
aniso8601==9.0.1
attrs==20.3.0
bcrypt==3.2.0
blinker==1.4
CacheControl==0.12.6
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-JWT-Extended==4.1.0
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
grpcio==1.36.1
gunicorn==20.0.4
httplib2==0.19.1
idna==2.10
importlib-metadata==3.10.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-core==4.7.1
jwcrypto==0.8
MarkupSafe==1.1.1
mongoengine==0.23.0
msgpack==1.0.2
nbformat==5.1.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
pyparsing==2.4.7
pyrsistent==0.17.3
python-dotenv==0.15.0
python-jwt==3.3.0
pytz==2021.1
pywin32==300
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.15.0
sseclient==0.0.27
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.1
